test(ssh): consolidate shared test helpers into helpers_test.go
Move helpers used across SSH test files into a single helpers_test.go, and
reorder each test file tests-first with file-local helpers at the bottom.
